requirements.txt
# Core dependencies
# Python 3.14+ required (already installed - cannot be installed via pip)
requests==2.31.0
pycurl==7.45.7
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pillow==11.3.0
tqdm==4.67.1
cryptography==45.0.5
standard-imghdr==3.10.14
certifi>=2024.0.0
keyring>=25.0.0
markdown>=3.4.0

# Windows-specific
colorama==0.4.6
pywin32-ctypes==0.2.3
winregistry==2.1.1
